'game' (Iteration) changed.
'game' now written in a more functional way, using
std::forward_list as return type for Iteration.

but: there is a design bug in c++11 for std::forward_list,
so empty braces initialization does not work, see:
http://stackoverflow.com/questions/24510067/why-does-return-not-apply-to-stdforward-list
, so this is now a c++14 application.

but: the STL implementation for my compiler (gcc) has
not fixed the bug yet (but it seems to be fixed soon).
hence, I added a dirty fix in IterationStack: #define _